requirement.txt
fastapi 
uvicorn 
python-dotenv 
pymongo 
motor 
requests 
langchain 
langchain-openai 
pydantic 
streamlit 
streamlit-chat 
loguru
